refactor(worker): 将 Worker 实现移至 client 子目录
