+++ backend/internal/worker/client/ws_client.go
@@ -1,4 +1,4 @@
-package worker
+package client
 
 import (
 	"context"
